Reviews/requirements.txt
alabaster==1.0.0
babel==2.16.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coverage==7.6.7
cryptography==43.0.3
Deprecated==1.2.15
docutils==0.21.2
Flask==3.1.0
Flask-Limiter==3.8.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
gprof2dot==2024.6.6
greenlet==3.1.1
idna==3.10
imagesize==1.4.1
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
memory-profiler==0.61.0
ordered-set==4.1.0
packaging==24.2
pluggy==1.5.0
psutil==6.1.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.0
PyMySQL==1.1.1
pytest==8.3.3
pytest-cov==6.0.0
pytest-profiling==1.7.0
requests==2.32.3
rich==13.9.4
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0